Implemented more drop files & reworked the dropfile system.
Should cover almost any drop files.

+++ crates/icy_board/src/menu_runner/pcb/e_enter_message.rs
@@ -64,7 +64,7 @@ impl PcbBoardCommand {
         };
 
         self.write_message(
-            self.state.session.current_conference_number,
+            self.state.session.current_conference_number as i32,
             area as i32,
             &to,
             &subject,
